+++ src/main/java/gift/Application.java
@@ -1,6 +1,6 @@
 package gift;
 
-import gift.user.oauth.KakaoProperties;
+import config.KakaoProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
