+++ src/main/java/gift/user/model/dto/KakaoAuthToken.java
@@ -1,4 +1,4 @@
-package gift.user.oauth;
+package gift.user.model.dto;
 
 public record KakaoAuthToken(String accessToken, String refreshToken) {
 }
